requirements.txt
Flask
Flask
Pillow
piexif
exifread
